Fix the focus going to the DayPicker instead of the Keyboard Shortcuts Panel when opened while focus is on the input.
Previously we were setting focus on the DayPicker and requiring that it have focus before we could show the Keyboard Shortcuts Panel so the CalendarDay would end up getting the focus instead of the close button inside the Keyboard Shortcuts Panel.  Now only the Panel should be getting the focus.